Http Server: Implicitly call json formatting for json handlers
Most of the API method return json type object. This patch remove the
need to explicitly call the json formatter when returning the result.

To do so a wrapper for the return type was added, it uses autoboxing to
translate return types to their json formatted representation.

The set_handler method with default return type, is set to return json
and to accept lambda expression that return the wrapper.

After the change it is possible to call:
getOSversion.set_handler([](const_req req)
    {
        return osv::version();
    });

The return string will be translated to json string.
No need to call the formatter anymore and no need to set the return type
to json.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>