cloud-init: client: make connection time out after 1 second
The default timeout seems too long. 1 second seems more than enough to
ping the local meta data service.

Note on the implementation: There really seems to be no way to set
timeout using boost asio sync api. The author recommends using async
API and a timer to achieve that. This makes the code more bloated but
if we want to stick with boost asio we don't have a choice according
to my research.

Signed-off-by: Tomasz Grabiec <tgrabiec@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>