net: simplify inpcb allocation
Remove inpcbinfo::ipi_zone, replacing it by a normal new/delete.  This reduces
indirections and makes it easier to modify the stack.

The call to uma_zone_set_max() is lost, but since it was a no-op, there
is no change in functionality.  If desired, it can be restored using
something like

  -new inpcb(...);
  +new (resticted_allocator) inpcb(...);

where restricted_allocator fails if we overflow the socket limit.

Signed-off-by: Avi Kivity <avi@cloudius-systems.com>