core: request termination of detached commands before proceeding to shutdown
Letting commands run while we shut down may result in page faults due
to DSOs getting unmapped while in use. This is often the case with
httpserver. This change will make OSv trigger termination of detached
commands before it shuts down. The callback is supposed to cause the
application to return from its main() in a timely manner.