core: integrate osv::run with osv::application
osv::application models a process more accurately than osv::run and
hence it should be used instead in most of the places which currently
use osv::run. This change moves osv::run() logic into osv::application
and makes the run() variants create an application.

Because users of run() are expected to get informed whether
application was run or not we need to change the API so that
application() will thrown a launch_error, instead of aborting, which
can be then caught by interested parties. This has a nice feature of
propagating more details about why we can't run the application.