revert "make image=", add new "make modules="
This is a compromise between some people's insistance (see issue #390)
that we shouldn't have modules added by default, and other people's
just-as-vigorous insistance that some modules, like httpserver, are
essential in every image.

What this patch does is to:

  1. return "make image=..." to how it used to be, i.e., only modules
     explicitly listed on this list, and their dependencies, are added
     to the image, and no "default" modules are added.

  2. A new "make modules=..." behaves like "make image=" does before
     this patch, i.e., it builds an image using both the listed modules,
     and the default modules. "make modules=x,-httpserver" will still work,
     if anyone wants to do this.

Internally, what this patch does is to add to the module list syntax a
prefix "!" which signifies we want this list exactly, without defaults.
The "make modules" doesn't add this prefix, while "make image" does
(i.e., "make image=x" is equivalent to "make modules=!x").

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>