Add rwlock read-lock benchmark to tests/misc-mutex
This patchs adds to tests/misc-mutex measurements of rwlock's reader
lock.

As I explain below, this is both a benchmark of rwlock itself (showing
the sad state it is in), and another benchmark for contended mutexes.

Ideally, in a read-mostly workload, we'd like to see a very quick
reader lock and excellent scalability (one CPU reading shouldn't bother
another CPU reading). But what we measure in this benchmark is the
complete opposite :-( Instead of a read-lock cycle taking a couple of
nanoseconds we measure uncontended read lock cycle at 50ns, twice slower
than a mutex (understandable, given rwlock's implementation); When two
threads on two different cpus want to read concurrently, the lock cycle,
jumps to 1800ns.

The reason for this slowness is twofold: First, we need a better, scalable,
rwlock implementation ;-) But second, taking as a given that we do have the
current rwlock implementation, which translates concurrent read access to a
mutex contention, we could benefit from better performance when the mutex
is contended. Adding spinning to the mutex dropped the 1800 ns
read-lock/unlock cycle to just 110 ns.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>