Http Server: Support optional Access-Control-Allow-Origin header
When the GUI and API are not placed on the same server it is useful to
have the Access-Control-Allow-Origin header set to *.

This adds a command line flag option to set the header.

The server now passes its configuration to the request handler.
When the flag is set, the request adds the access control header.

Reviewed-by: Tomasz Grabiec <tgrabiec@gmail.com>
Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>