Http server: Expose ifconfig infromation in the API
This adds the equivalent of the ifconfig information to the API.
/network/ifconfig will return a list of all the interfaces with their
configuration and data.
To check a specific interface (ie. lo0) do:
/network/ifconfig/lo0

Interfaces information returns with timestamp. Currently all interfaces
in a single call have the same timestamp, we assume that the
difference is neglegtable, but API support specific
timestamp per interface.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>