upload_manifest: be more forgiving with encoding errors
There are situations in which OSv may print garbage to the console. In particular,
I am printing the contents of a stream for debugging purposes, and while it is sane
in the case I am handling, the build process ends up hitting the same trigger - and at
that point, it contains garbage. We fail, because we can't decode the characters to the
default encoding (ascii).

This patch changes the error mode of python's decode method from 'strict' (default) to
'replace', which means "print something". Since we're touching the parameters anyway,
the encoding is also changed to UTF-8.

Reviewed-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>