modules: introduce run_on_init() declaration
api.run_on_init(cmdline) means that cmdline should be executed at init
time as if a file was put inside /init directory with '30-'
prefix.

The order in which commands declared this way will run depends on the
order in which modules depend on each other. Dependant modules will
run before modules which depend on them which means automatic priority
assignment.

The motivation for this change is to allow "run on init" behavior to
be optional, depending on the chosen run configuration. For example
some times we want HTTP server to run as daemon and some times as
standalone app. It's simplest to reuse an already present run
configuration mechanism for that.