Fix errors in "make clean"
Currently, "make clean" prints errors like this:

   /bin/sh: ./scripts/osv-version.sh: No such file or directory

The problem was that apps/rpm/Makefile uses $(OSV_BASE)/scripts/osv-version.sh
and it was passed OSV_BASE=., but the current directory was changed.
So instead of passing ".", pass the full path name of the current directory,
and the error messages are gone.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>