musl: take trigonometric functions from upstream musl
Take a bunch of trigonometric functions from the upstream musl directory,
dropping our own copy.

The upstream files are *not* identical to the ones we had, but our files
were simply taken from a slightly different version, and had no
modifications, so we have no reason to keep the specific version we had.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@cloudius-systems.com>