top.py: also print total idle time
on a 4-cpu guest, top.py currently prints something like this:

	idle:  94%  75%  92%  92%

Where the first CPU is 94% idle, and the second CPU is 75% idle,
and so on. This patch adds another statistic, the sum of the CPUs'
idle times:

	idle:  94%  75%  92%  92%  (total 353%)

The number "353%" means that 353% of the strength of one CPU are idle.
This is often a useful statistic - on a fully loaded guest, we want to
see this get near to 0%.

The reason why we print the total idle percentage (353%) and not the
average (353/4 = 88%) is to keep all percentages in top.py consistent
in that they are of a single CPU, not of the total CPUs. If we start
printing both types of percentages on the same screen, it will be too
confusing.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>