Move tracepoint_count functions into the kernel
Before this patch, trace-count.hh was purly a header file, and when
we used it in a shared object (such as httpserver.so), the code of
the tracepoint hit() function ended up part of the shared object.

Because the hit() function often runs in preemption-disabled context,
this caused a bunch of problems. One problem we already worked around
was that on the first call, it might call a function for the first
time and cause a (possibly sleeping) symbol resolution. But another
bug we never solved was that pages from the shared object might be
evicted, in which case suddenly hit() will cause a page fault.

So this patch moves the hit() implementation out of line, from
trace-count.hh to a new trace-count.cc, i.e., into the kernel's code.
While we're at it, I moved the other inline functions out of line too.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>