tests: gtod
I have seen gettimeofday - and the pvclock functions in general - show up
quite heavily in some profiles (some 4 - 5 %).

Although this is of course normal an expected for a system that is doing a lot
of time operations, I wanted to make sure that we are not going anything too
wasteful in there.

I wrote this simple benchmark to attest that, since we seem to have none so
far. The good news is that we are quite faster than a CentOS guest, albeit
2 x slower than the host (hard to avoid that).

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>