console: fix support for discarding or keeping input
When switching the tty to raw mode, we have two options about what to
do with the unread input, including perhaps an partially edited line:

1. Keep the the partially edited line and make it available for reading
   imediately (it can never be edited again, even if we later go back to
   cooked mode).
   This happens with TCSETS ioctl, or tcsetattr's TCSANOW option.

2. Discard all unread input, including a partial edited line.
   This happens with TCSETSF ioctl, or tcsetattr's TCSAFLUSH option.

Unfortunately, we supported neither of these two options correctly...
Instead, what we did in both cases was to "hide" the partially edited
line for the duration of the raw mode, and "bring it back" when going
back to cooked mode. This is *not* how Linux behaves, and caused problems
in one application which tried to switch to raw mode after input was
already available.

So this patch fixes these two cases to each do the right thing.
We add to the line-discipling object two new operations - take_pending_input
(copying the partially edited line into the ready-to-read buffer),
and discard_pending_input (discarding both partially edited line and the
ready-to-read buffer). We then call these two new operations as appropriate
from the ioctl operations of the console device.

Fixes #461

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>