Lua: Fix build flags and libraries
The flag LUA_USE_LINUX implicitly sets LUA_USE_READLINE, which then makes the
build require 'readline.h'. We don't use readline, and readline headers might
not be available on the machine it's built with. LUA_USE_LINUX is a group of
other flags. Include this group, but exclude LUA_USE_READLINE.

LuaRocks requires POPEN support in Lua, so this flag was added as well.

Remove -ledit from the build as it is not needed.

I also took the chance to refactor the module Makefile.

Signed-off-by: Or Cohen <orc@fewbytes.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>