net: Remove poll_wake in sofree
The call chain is:

   fp->stop_polls -> fp->close -> soclose -> sorele -> sofree ->
   poll_wake -> fhold -> fdrop

We set so->fp to NULL in soclose, we do nothing in poll_wake if so->fp
is NULL. So no need to call poll_wake in sofree.

Besides, before we call fp->close, fp->stop_polls will stop all the
polls, so the poll list should be empty anyway.

Signed-off-by: Asias He <asias@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>