zfs: initialize znode's z_vnode field
ZFS on OSv initializes a znode without necessarily having to associate
a vnode, e.g. on certain operations that don't involve looking up the
target node, such as unlink. Then, z_vnode should be initialized to
null to make it explicit that the znode has no vnode associated, and
it's also the correct thing to do.

Signed-off-by: Raphael S. Carvalho <raphaelsc@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>