osv::reboot(): don't halt on failure
osv::reboot()'s documentation says that if it fails to reboot it may
return. So don't call halt(). This halt() was only added as a stub for ARM,
so leave it only for the ARM case (and add a "FIXME" comment, as this
should be replaced by a real reboot).

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>