aarch64: gic: set target mask to zero for the uniproc
in commit ad8466ba18d24c72c96c4d78d173f5042974f5f7 we removed
an abort and changed it to a warning when we cannot read our
own cpu mask target from the GIC, as zero is allowed as a
special case for uniprocessor.
However, we logically still need to set the cpu target mask
to zero in our gic driver state.

Signed-off-by: Claudio Fontana <claudio.fontana@huawei.com>
Cc: Geyslan G. Bem <geyslan@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>