cpiod: respect original mode of files
modes are usually not very interesting for us, so we have been able to live
without it just fine. When uploading files to the guest, we only set the type
bits, and left the rest blank.

However, there are some programs (like MySQL), that will conduct tests on the
permission flags of some of their configuration files. The poor guys don't know
they are running on OSv, and will therefore abort when seeing what they believe
to be a insecure setting.

After this patch, all the modes are respected, and if one needs to upload files
with specific modes, it's just a matter of guaranteeing that the host file is
set correctly.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@cloudius-systems.com>