build: do not use 'private' in build.mk
Commit a4b52857d38427c784adeee48a3a513f89fd3450 introduced 'private'
keyword to solve a dependencies issue caused by the inheritance of
target-specific variable. The keyword however is only available since
make 3.82.

This patch removes the '-lstatic-thread-variable' option. Since we've
already added the library as one of the target's prerequisites and
compiling the target with '$^', that option is not needed.

Fixes #405.

Reviewed-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Zifei Tong <zifeitong@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>