aarch64: dtb: read amount of memory from device tree
start making use of device trees for reading the amount of memory.

FDT (Flattened Device Tree) is made available to the guest as a
DTB (Device Tree Blob).

Reading of the dtb requires some reshuffling.

The DTB is moved from its initial location to its final place in
physical memory at the location OSV_KERNEL_BASE, overwriting
the preboot code (preboot.S), and also overwriting the command
line and the like.
Therefore we need to parse the command line before the DTB move.

Some adaptation of the memory setup code is also needed,
to account for a dynamic amount of memory.

Signed-off-by: Claudio Fontana <claudio.fontana@huawei.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>