top.py: fix no-cpu case
A sched::thread object can be created and not start()ed.
We often have two such unstarted threads created for itimer when this
feature is not used by the application.

Such a thread has no CPU, and currently top.py prints is as an ugly
"4294967295" for these threads which ruins the nice column alignment.
This patch changes this case to print "-" as the CPU instead.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
[ penberg: Use hexadecimal constant ]
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>