Http Server: Handling of a type conversion in complex types
The Json objects allow auto type conversion, based on fields name. To
handle situation where translated type contains other structures the
setting functionality was modified.

First the basic json element assignment has an additional template
implementation allowing setting it's value with any type the compiler
can assign. i.e. if field is of type A, then the json_element<A> can be
assigned any type B if A can be assigned from B.

Second, the set template method that was defined in a json object to
copy all it's fields was rename to be a template assignment operator.

So json object A can be assigned from any type B if B has all the fields
A has and each of the fields in A can be assigned from the field in B
with the same name.

If a field in A is a structure, the equivalent field in B should be a
structure with the same above limitation.

Verified-by: Vlad Zolotarov <vladz@cloudius-systems.com>
Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>