top.py: make number of lines a parameter
top.py currently hard-coded the number of top threads shown to 20.
This patch adds a parameter, "-l" (or "--lines") allowing to change
this default. This is useful for example if you intend to dump the
output of top.py to a file, in which case there is no reason to limit
it to 20 threads, and you can just as easily ask for 500.

Reviewed-by: Tomasz Grabiec <tgrabiec@gmail.com>
Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>