Drop write_lock while retrieving ARC buffer
ZFS code may sleep if data the page cache is asking for is not present
in memory. Sleeping while holding a lock will prevent other threads from
processing page fault. This patch drops write_lock while calling into
ZFS, but after lock is acquired back we need to recheck that page was
not instantiated in a write cache. If it was write page cache needs to
be mapped instead of ARC cache. The patch achieves that by returning
to a faulting thread without mapping anything at which point fault is
retried and write cache page is mapped.

Signed-off-by: Gleb Natapov <gleb@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@cloudius-systems.com>