java: Mechanism to auto-load resident JVMTI agents iff included in image
Will make java launcher add any JVMTI agent (i.e. .so) present in
'/usr/lib/jvm/agents/autoload/'.

Since dickering with actual appliance command lines is error prone,
and having a feature (agent) included or not in the image should
be distinction enough, picking up extensions this way feels ok.

Purpose of course to be able to have optional java symbol agent
loaded into the jvm if the image is built with it.

Reviewed-by: Tomasz Grabiec <tgrabiec@gmail.com>
Signed-off-by: Calle Wilund <calle@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>