Instruct people to use scripts/setup.py for Fedora prerequisites
The scripts/setup.py script is there for installing prerequisites. It's
known to work on Fedora so drop the duplicate list of prerequisite
packages from README.md and just tell people to use scripts/setup.py.

We really should do the same thing for other operating systems as well
but that should be done by people who have access to them.

Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>